requirements.txt
mcp[cli]
python-dotenv
requests
anthropic
python-telegram-bot
discord.py
